Move physics object cleanup into MScene.
Now when an object is removed from the scene, we check if it has a corresponding physics object as well. If so, we remove that too.

This could cause issues if someone else is manually cleaning up the physics context before calling MScene::deleteObject, but I couldn't find any instances of that occuring.

Tested by running the game and the editor. I verified that deleting objects works via the editor.