Add OnCollisionBehavior.
The OnCollisionBehavior runs a lua function when the parent object and another designated object collide. The lua function does not receive any arguments; this will be addressed in a later patch.

There's a known issue right now where adding this custom behavior to the scene causes the player position to not reset in the editor when the game exits. Until this bug is tracked down, I won't change the scene file.

This change will cause the VS9 build to break. Someone will need to add
  src/on-collision-behavior.cc
  src/on-collision-behavior.h
to the Maratis and MaratisEditor projects.